pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "notes-manager"
version = "1.0.0"
description = "Obsidian vault management skill for Claude Code"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "click>=8.0",
    "rich>=13.0",
]

[project.optional-dependencies]
nlp = [
    "scikit-learn>=1.3",
    "nltk>=3.8",
    "networkx>=3.1",
]

[project.scripts]
notes-manager = "src.main:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
